bug fixes, adding auth jwt, & some validation

+++ src/controller/products.js
@@ -15,60 +15,83 @@ const {
 const getAllProducts = async (req, res) => {
   const urlQueries = req.query;
   const result = await selectAllProducts(urlQueries);
-  jsonResponse(res, result);
+  return jsonResponse(res, result);
 };
 
 const getProduct = async (req, res) => {
   const { id } = req.params;
   const result = await selectDataProduct(id);
-  jsonResponse(res, result[0]);
+  return jsonResponse(res, result);
 };
 
 const postProduct = async (req, res) => {
   try {
     let data = req.body;
-    data.image = req.file.path;
+    data.image = req.file ? req.file.path : '';
     const query = await insertDataProduct(data);
     const id = query.insertId;
     const result = {
       id,
       ...data
     };
-    jsonResponse(res, result);
+    return jsonResponse(res, result);
   } catch(error) {
-    jsonError(res, errorBadRequest);
-    throw new Error(error);
+    return jsonError(res, errorBadRequest);
   }
 };
 
 const putProduct = async (req, res) => {
   try {
     let data = req.body;
-    data.image = req.file.path;
     const { id } = req.params;
-    await updateDataProduct(data, id);
-    const result = {
-      id,
-      ...data
-    };
-    jsonResponse(res, result);
+    if(req.file) {
+      data.image = req.file.path;
+    }
+    await selectDataProduct(id).then(async product => {
+      if(product.id) {
+        await updateDataProduct(data, id);
+        const result = {
+          id,
+          ...data
+        };
+        return jsonResponse(res, result);
+      } else {
+        return jsonError(res, errorBadRequest);
+      }
+    }).catch(() =>  {
+      const error = {
+        code: 400,
+        message: `Data product does not exist!`
+      };
+      return jsonError(res, error);
+    });
   } catch(error) {
-    jsonError(res, errorBadRequest);
-    throw new Error(error);
+    return jsonError(res, errorBadRequest);
   }
 };
 
 const deleteProduct = async (req, res) => {
   try {
     const { id } = req.params;
-    await deleteDataProduct(id);
-    const result = {
-      id
-    };
-    jsonResponse(res, result);
+    await selectDataProduct(id).then(async product => {
+      if(product.id) {
+        await deleteDataProduct(id);
+        const result = {
+          ...product
+        };
+        return jsonResponse(res, result);
+      } else {
+        return jsonError(res, errorBadRequest);
+      }
+    }).catch(() =>  {
+      const error = {
+        code: 400,
+        message: `Data product does not exist!`
+      };
+      return jsonError(res, error);
+    });
   } catch(error) {
-    jsonError(res, errorBadRequest);
-    throw new Error(error);
+    return jsonError(res, errorBadRequest);
   }
 };
